@@ -1,27 +1,27 @@
 {
-  "_from": "mime-db@1.45.0",
-  "_id": "mime-db@1.45.0",
+  "_from": "mime-db@1.46.0",
+  "_id": "mime-db@1.46.0",
   "_inBundle": false,
-  "_integrity": "sha512-CkqLUxUk15hofLoLyljJSrukZi8mAtgd+yE5uO4tqRZsdsAJKv0O+rFMhVDRJgozy+yG6md5KwuXhD4ocIoP+w==",
+  "_integrity": "sha512-svXaP8UQRZ5K7or+ZmfNhg2xX3yKDMUzqadsSqi4NCH/KomcH75MAMYAGVlvXn4+b/xOPhS3I2uHKRUzvjY7BQ==",
   "_location": "/mime-db",
   "_phantomChildren": {},
   "_requested": {
     "type": "version",
     "registry": true,
-    "raw": "mime-db@1.45.0",
+    "raw": "mime-db@1.46.0",
     "name": "mime-db",
     "escapedName": "mime-db",
-    "rawSpec": "1.45.0",
+    "rawSpec": "1.46.0",
     "saveSpec": null,
-    "fetchSpec": "1.45.0"
+    "fetchSpec": "1.46.0"
   },
   "_requiredBy": [
     "/mime-types"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-db/-/mime-db-1.45.0.tgz",
-  "_shasum": "cceeda21ccd7c3a745eba2decd55d4b73e7879ea",
-  "_spec": "mime-db@1.45.0",
-  "_where": "/Users/dilaraavci/Sites/ToDoListe/backend/node_modules/mime-types",
+  "_resolved": "https://registry.npmjs.org/mime-db/-/mime-db-1.46.0.tgz",
+  "_shasum": "6267748a7f799594de3cbc8cde91def349661cee",
+  "_spec": "mime-db@1.46.0",
+  "_where": "/Users/dilaraavci/Sites/Semesteraufgabe/backend/node_modules/mime-types",
   "bugs": {
     "url": "https://github.com/jshttp/mime-db/issues"
   },
@@ -48,16 +48,16 @@
     "bluebird": "3.7.2",
     "co": "4.6.0",
     "cogent": "1.0.1",
-    "csv-parse": "4.12.0",
-    "eslint": "7.9.0",
-    "eslint-config-standard": "14.1.1",
-    "eslint-plugin-import": "2.22.0",
+    "csv-parse": "4.15.1",
+    "eslint": "7.20.0",
+    "eslint-config-standard": "15.0.1",
+    "eslint-plugin-import": "2.22.1",
     "eslint-plugin-markdown": "1.0.2",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "4.2.1",
-    "eslint-plugin-standard": "4.0.1",
+    "eslint-plugin-promise": "4.3.1",
+    "eslint-plugin-standard": "4.1.0",
     "gnode": "0.1.2",
-    "mocha": "8.1.3",
+    "mocha": "8.3.0",
     "nyc": "15.1.0",
     "raw-body": "2.4.1",
     "stream-to-array": "2.3.0"
@@ -93,10 +93,10 @@
     "fetch": "node scripts/fetch-apache && gnode scripts/fetch-iana && node scripts/fetch-nginx",
     "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --reporter spec --bail --check-leaks test/",
+    "test-ci": "nyc --reporter=lcov --reporter=text npm test",
     "test-cov": "nyc --reporter=html --reporter=text npm test",
-    "test-travis": "nyc --reporter=text npm test",
     "update": "npm run fetch && npm run build",
     "version": "node scripts/version-history.js && git add HISTORY.md"
   },
-  "version": "1.45.0"
+  "version": "1.46.0"
 }
