@@ -1,8 +1,8 @@
 {
   "_from": "mime-types@~2.1.24",
-  "_id": "mime-types@2.1.28",
+  "_id": "mime-types@2.1.29",
   "_inBundle": false,
-  "_integrity": "sha512-0TO2yJ5YHYr7M2zzT7gDU1tbwHxEUWBCLt0lscSNpcdAfFyJOVEpRYNS7EXVcTLNj/25QO8gulHC5JtTzSE2UQ==",
+  "_integrity": "sha512-Y/jMt/S5sR9OaqteJtslsFZKWOIIqMACsJSiHghlCAyhf7jfVYjKBmLiX8OgpWeW+fjJ2b+Az69aPFPkUOY6xQ==",
   "_location": "/mime-types",
   "_phantomChildren": {},
   "_requested": {
@@ -19,10 +19,10 @@
     "/accepts",
     "/type-is"
   ],
-  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.28.tgz",
-  "_shasum": "1160c4757eab2c5363888e005273ecf79d2a0ecd",
+  "_resolved": "https://registry.npmjs.org/mime-types/-/mime-types-2.1.29.tgz",
+  "_shasum": "1d4ab77da64b91f5f72489df29236563754bb1b2",
   "_spec": "mime-types@~2.1.24",
-  "_where": "/Users/dilaraavci/Sites/ToDoListe/backend/node_modules/type-is",
+  "_where": "/Users/dilaraavci/Sites/Semesteraufgabe/backend/node_modules/type-is",
   "bugs": {
     "url": "https://github.com/jshttp/mime-types/issues"
   },
@@ -44,19 +44,19 @@
     }
   ],
   "dependencies": {
-    "mime-db": "1.45.0"
+    "mime-db": "1.46.0"
   },
   "deprecated": false,
   "description": "The ultimate javascript content-type utility.",
   "devDependencies": {
-    "eslint": "7.17.0",
+    "eslint": "7.20.0",
     "eslint-config-standard": "14.1.1",
     "eslint-plugin-import": "2.22.1",
-    "eslint-plugin-markdown": "1.0.2",
+    "eslint-plugin-markdown": "2.0.0",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "4.2.1",
+    "eslint-plugin-promise": "4.3.1",
     "eslint-plugin-standard": "4.1.0",
-    "mocha": "8.2.1",
+    "mocha": "8.3.0",
     "nyc": "15.1.0"
   },
   "engines": {
@@ -79,10 +79,10 @@
     "url": "git+https://github.com/jshttp/mime-types.git"
   },
   "scripts": {
-    "lint": "eslint --plugin markdown --ext js,md .",
+    "lint": "eslint .",
     "test": "mocha --reporter spec test/test.js",
     "test-ci": "nyc --reporter=lcov --reporter=text npm test",
     "test-cov": "nyc --reporter=html --reporter=text npm test"
   },
-  "version": "2.1.28"
+  "version": "2.1.29"
 }
