@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { ApiService } from './../services/api.service';
 import { AuthService } from './../services/auth.service';
+import {User} from "../models";
+import {AccountService} from "../shared/account.service";
 
 @Component({
   selector: 'app-dashboard',
@@ -9,12 +11,16 @@ import { AuthService } from './../services/auth.service';
 })
 export class DashboardComponent implements OnInit {
 
+  user: User;
   isLogin = false;
 
   constructor(
     private api: ApiService,
     private auth: AuthService,
-  ) { }
+    private accountService: AccountService
+  ) {
+    this.accountService.user.subscribe(x => this.user = x);
+  }
 
   ngOnInit(): void {
   }
